+++ test/schema.js
@@ -118,6 +118,30 @@ metatests.test('Schema: nested json, lost field', (test) => {
   test.end();
 });
 
+metatests.test('Schema: optional nested struct', (test) => {
+  const definition = {
+    struct: {
+      required: false,
+      json: {
+        field: 'string',
+      },
+    },
+  };
+  const schema = Schema.from(definition);
+
+  const obj1 = {};
+  test.strictSame(schema.check(obj1).valid, true);
+
+  const obj2 = {
+    struct: {
+      field: 'value',
+    },
+  };
+  test.strictSame(schema.check(obj2).valid, true);
+
+  test.end();
+});
+
 metatests.test('Schema: shorthand', (test) => {
   const definition = {
     field1: {